Add files via upload
Predict if a customer will purchase a product or not based on his/her User ID, Gender, Age, and Estimated Salary. Since gender and user id do not matter in such predictions, they are not included in the feature set passed to the DecisionTree classifier for prediction of purchase of item.