Add files via upload
Mean shift clustering aims to discover “blobs” in a smooth density of samples. It is a centroid-based algorithm, which works by updating candidates for centroids to be the mean of the points within a given region. These candidates are then filtered in a post-processing stage to eliminate near-duplicates to form the final set of centroids. 
Comparison of Mean-Shift and K-means clustering:
Disadvantages of Mean-Shift clustering:
1) O(T*n*log(n)) in lower dimensions, with n the number of samples and T the number of points
2) O(T*n^2) in higher dimensions